Move tests around to appease `tidy`
